app/requirements.txt
huggingface-hub
pandas
python-dotenv
requests
celery
flask
redis